不要なtypes削除
